chore: fix incorrect relative path
The compile.sh is called by the runall.sh in the directory according to
the benchmark we run; however, it assumes that it's always called under
the project root, leading to incorrect relative path, causing
compilation error.
The path to the project root is now taken by the script as an argument.
Additionally, description are added for each arguments in the help
message.